chore: move dependency repo to new folder
