+++ api/dependencies/registration.go
@@ -3,11 +3,12 @@ package dependencies
 import (
 	"github.com/neo4j/neo4j-go-driver/v5/neo4j"
 	"net/http"
-	"service-dependency-api/api/dependencies/internal/dependencyRepository"
+	"service-dependency-api/neo4jRepositories/dependencyRepository"
+	"service-dependency-api/repositories"
 )
 
 type ServiceCallsHandler struct {
-	Repository dependencyRepository.DependencyRepository
+	Repository repositories.DependencyRepository
 }
 
 func (s *ServiceCallsHandler) Register(mux *http.ServeMux) {
@@ -23,10 +24,8 @@ func (s *ServiceCallsHandler) Register(mux *http.ServeMux) {
 }
 
 func Register(mux *http.ServeMux, driver *neo4j.DriverWithContext) {
-	repo := dependencyRepository.New(*driver)
-
 	handler := ServiceCallsHandler{
-		Repository: repo,
+		Repository: dependencyRepository.New(*driver),
 	}
 	handler.Register(mux)
