+++ repositories/dependencies.go
@@ -1,0 +1,17 @@
+package repositories
+
+import "errors"
+
+type Dependency struct {
+	Id          string `json:"id"`
+	Version     string `json:"version,omitempty"`
+	Name        string `json:"name,omitempty"`
+	ServiceType string `json:"type,omitempty"`
+}
+
+func (d *Dependency) Validate() error {
+	if d.Id == "" {
+		return errors.New("dependency id is required")
+	}
+	return nil
+}
